requirements.txt
numpy
pandas
torch
torchvision
Pillow
tqdm
deepface
dlib
matplotlib
mediapipe